+++ src/VkApp.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 //
 // Light Vulkan Graphics
-// Copyright (C) 2025 Dr. Nathanael John Inkson
+// Copyright (C) 2026 Dr. Nathanael John Inkson
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Lesser General Public License as published
@@ -4909,58 +4909,6 @@ std::shared_ptr<Texture> VkApp::createTextureFromEmbedded(const EmbeddedTextureD
 		markLightingDirty();
 	}
 
-	void VkApp::addObject(lightGraphics::pObject *newObject)
-	{
-		if (!newObject)
-		{
-			throw std::invalid_argument("addObject: object pointer is null");
-		}
-
-		_objects_.push_back(*newObject);
-
-		// Initialize dirty tracking for new object
-		dirtyObjects_.push_back(true);
-		objectModelMatrixOverrides_.push_back(std::nullopt);
-		instanceDataDirty_ = true;
-
-		if (sceneFinalized_)
-		{
-			updateInstanceData(); // Update rendering data only if scene is finalized
-		}
-	}
-
-	void VkApp::addObject(const lightGraphics::pObject& obj)
-	{
-		_objects_.push_back(obj);
-
-		// Initialize dirty tracking for new object
-		dirtyObjects_.push_back(true);
-		objectModelMatrixOverrides_.push_back(std::nullopt);
-		instanceDataDirty_ = true;
-
-		if (sceneFinalized_)
-		{
-			updateInstanceData();
-		}
-	}
-
-	void VkApp::addObject(lightGraphics::ShapeType type, const glm::vec3& position,
-						const glm::vec3& size, const glm::vec4& color,
-						const glm::quat& rotation, const std::string& name, float mass)
-	{
-		_objects_.emplace_back(type, position, size, color, rotation, name, mass);
-
-		// Initialize dirty tracking for new object
-		dirtyObjects_.push_back(true);
-		objectModelMatrixOverrides_.push_back(std::nullopt);
-		instanceDataDirty_ = true;
-
-		if (sceneFinalized_)
-		{
-			updateInstanceData();
-		}
-	}
-
 	size_t VkApp::addRiggedObject(const std::shared_ptr<RiggedObject>& riggedObject)
 	{
 		if (!riggedObject)
@@ -5131,66 +5079,6 @@ std::shared_ptr<Texture> VkApp::createTextureFromEmbedded(const EmbeddedTextureD
 		sceneGraph_->onRiggedObjectRemoved(index);
 	}
 
-	glm::mat4 VkApp::getObjectModelMatrix(size_t index) const
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("getObjectModelMatrix", index, _objects_.size()));
-		}
-
-		if (index < objectModelMatrixOverrides_.size() && objectModelMatrixOverrides_[index])
-		{
-			return *objectModelMatrixOverrides_[index];
-		}
-
-		const auto& obj = _objects_[index];
-		const glm::mat4 translation = glm::translate(glm::mat4(1.0f), obj.getPosition());
-		const glm::mat4 rotation = glm::mat4_cast(obj.getRotation());
-		const glm::mat4 scale = glm::scale(glm::mat4(1.0f), obj.getSize());
-		return translation * rotation * scale;
-	}
-
-	void VkApp::setObjectModelMatrixOverride(size_t index, const glm::mat4& model)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("setObjectModelMatrixOverride", index, _objects_.size()));
-		}
-
-		if (objectModelMatrixOverrides_.size() < _objects_.size())
-		{
-			objectModelMatrixOverrides_.resize(_objects_.size());
-		}
-		objectModelMatrixOverrides_[index] = model;
-
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
-		else
-		{
-			instanceDataDirty_ = true;
-		}
-	}
-
-	void VkApp::clearObjectModelMatrixOverride(size_t index)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("clearObjectModelMatrixOverride", index, _objects_.size()));
-		}
-
-		clearObjectModelMatrixOverrideInternal(index);
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
-		else
-		{
-			instanceDataDirty_ = true;
-		}
-	}
-
 	void VkApp::setLightTransformMatrixOverride(size_t index, const glm::mat4& transform)
 	{
 		if (index >= lights_.size())
@@ -5236,101 +5124,6 @@ std::shared_ptr<Texture> VkApp::createTextureFromEmbedded(const EmbeddedTextureD
 			throw std::out_of_range(makeObjectIndexMessage("clearRiggedObjectTransformMatrixOverride", index, riggedInstances_.size()));
 		}
 		riggedInstances_[index].transformMatrixOverride.reset();
-	}
-
-	SceneGraph& VkApp::sceneGraph()
-	{
-		return *sceneGraph_;
-	}
-
-	const SceneGraph& VkApp::sceneGraph() const
-	{
-		return *sceneGraph_;
-	}
-
-	// Object update methods for physics simulation
-	void VkApp::setObjectPosition(size_t index, const glm::vec3& position)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("setObjectPosition", index, _objects_.size()));
-		}
-
-		_objects_[index].setPosition(position);
-		clearObjectModelMatrixOverrideInternal(index);
-		sceneGraph_->onObjectChanged(index);
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
-	}
-
-	void VkApp::setObjectScale(size_t index, const glm::vec3& scale)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("setObjectScale", index, _objects_.size()));
-		}
-
-		_objects_[index].setSize(scale);
-		clearObjectModelMatrixOverrideInternal(index);
-		sceneGraph_->onObjectChanged(index);
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
-	}
-
-	void VkApp::setObjectRotation(size_t index, const glm::quat& rotation)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("setObjectRotation", index, _objects_.size()));
-		}
-
-		_objects_[index].setRotation(rotation);
-		clearObjectModelMatrixOverrideInternal(index);
-		sceneGraph_->onObjectChanged(index);
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
-	}
-
-	void VkApp::setObjectColor(size_t index, const glm::vec4& color)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("setObjectColor", index, _objects_.size()));
-		}
-
-		_objects_[index].setColour(color);
-		sceneGraph_->onObjectChanged(index);
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
-	}
-
-	// Update multiple object properties at once for efficiency
-	void VkApp::updateObjectProperties(size_t index, const glm::vec3& position,
-									const glm::vec3& scale, const glm::quat& rotation)
-	{
-		if (index >= _objects_.size())
-		{
-			throw std::out_of_range(makeObjectIndexMessage("updateObjectProperties", index, _objects_.size()));
-		}
-
-		_objects_[index].setPosition(position);
-		_objects_[index].setSize(scale);
-		_objects_[index].setRotation(rotation);
-		clearObjectModelMatrixOverrideInternal(index);
-		sceneGraph_->onObjectChanged(index);
-
-		if (sceneFinalized_)
-		{
-			markObjectDirty(index);
-		}
 	}
 
 	// Physics update callback
@@ -5526,74 +5319,6 @@ std::shared_ptr<Texture> VkApp::createTextureFromEmbedded(const EmbeddedTextureD
 
 		addObject(lightGraphics::ShapeType::CYLINDER, cylinderCenter, size, color, rotation, name, mass);
 	}
-
-	void VkApp::removeObject(size_t index)
-	{
-		if (index < _objects_.size())
-		{
-			_objects_.erase(_objects_.begin() + index);
-			if (index < dirtyObjects_.size())
-			{
-				dirtyObjects_.erase(dirtyObjects_.begin() + index);
-			}
-			if (index < objectModelMatrixOverrides_.size())
-			{
-				objectModelMatrixOverrides_.erase(objectModelMatrixOverrides_.begin() + index);
-			}
-			if (index < instanceDataCache_.size())
-			{
-				instanceDataCache_.erase(instanceDataCache_.begin() + index);
-			}
-			sceneGraph_->onObjectRemoved(index);
-			instanceDataDirty_ = true;
-			if (sceneFinalized_)
-			{
-				updateInstanceData();
-			}
-		}
-	}
-
-	void VkApp::addHexahedral(const glm::vec3& position, const glm::vec3& size,
-							const glm::vec4& color,
-							const glm::quat& rotation,
-							const std::string& name,
-							float mass)
-	{
-		addObject(lightGraphics::ShapeType::HEX, position, size, color, rotation, name, mass);
-	}
-
-	void VkApp::clearObjects()
-	{
-		const size_t removedCount = _objects_.size();
-		_objects_.clear();
-		dirtyObjects_.clear();
-		objectModelMatrixOverrides_.clear();
-		instanceDataCache_.clear();
-		for (size_t i = 0; i < removedCount; ++i)
-		{
-			sceneGraph_->onObjectRemoved(0);
-		}
-		instanceDataDirty_ = true;
-		if (sceneFinalized_)
-		{
-			updateInstanceData();
-		}
-	}
-
-	void VkApp::updateObject(size_t index, const lightGraphics::pObject& obj)
-	{
-		if (index < _objects_.size())
-		{
-			_objects_[index] = obj;
-			clearObjectModelMatrixOverrideInternal(index);
-			sceneGraph_->onObjectChanged(index);
-			if (sceneFinalized_)
-			{
-				markObjectDirty(index);
-			}
-		}
-	}
-
 
 	// ==================== PERFORMANCE OPTIMIZATION METHODS ====================
 
@@ -5860,38 +5585,6 @@ std::shared_ptr<Texture> VkApp::createTextureFromEmbedded(const EmbeddedTextureD
 		}
 
 		instanceDataDirty_ = false;
-	}
-
-	// Batch update methods for high performance
-	void VkApp::updateObjectPositions(const std::vector<std::pair<size_t, glm::vec3>>& updates)
-	{
-		for (const auto& update : updates)
-		{
-			if (update.first < _objects_.size())
-			{
-				_objects_[update.first].setPosition(update.second);
-				clearObjectModelMatrixOverrideInternal(update.first);
-				sceneGraph_->onObjectChanged(update.first);
-				markObjectDirty(update.first);
-			}
-		}
-	}
-
-	void VkApp::updateObjectProperties(const std::vector<std::tuple<size_t, glm::vec3, glm::vec3, glm::quat>>& updates)
-	{
-		for (const auto& update : updates)
-		{
-			size_t index = std::get<0>(update);
-			if (index < _objects_.size())
-			{
-				_objects_[index].setPosition(std::get<1>(update));
-				_objects_[index].setSize(std::get<2>(update));
-				_objects_[index].setRotation(std::get<3>(update));
-				clearObjectModelMatrixOverrideInternal(index);
-				sceneGraph_->onObjectChanged(index);
-				markObjectDirty(index);
-			}
-		}
 	}
 
 	void VkApp::flushPendingUpdates()